requirements.txt
fastapi
uvicorn
pymongo
python-multipart
PyPDF2
python-docx
pydantic
